% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccs.R
\name{nccs}
\alias{nccs}
\title{Net Compare Cross Species of WGCNA}
\usage{
nccs(
  humangene,
  mousegene,
  humandata,
  mousedata,
  humannet,
  mousenet,
  humanmodcolor,
  mousemodcolor,
  k = 20,
  linecoloralpha = c(0.4, 0.4, 0.4, 0.4),
  mosttimes = 500,
  quitname = "result of nccsWGCNA",
  means = c("Human Transformational Adjacency Network",
    "Mouse Transformational Adjacency Network",
    "Human-Mouse Transformational Adjacency Network",
    "Mouse-Human Transformational Adjacency Network")
)
}
\arguments{
\item{humangene, mousegene}{must be a one-to-one corresponding gene list to be analyzed. We understand that when conducting single-cell comparative studies, in order to merge data, gene names in the dataset are often converted to homologous genes from another species. Therefore, a pre gene list geneidlist was provided. This is manually generated by us based on existing public information, providing basic information on human and mouse gene correspondence. The first column is the code, the second column is the human gene, and the third column is the mouse gene. Based on the gene list, we provide automatic conversion of human and mouse genes; When you need similar name conversions compared to other species, please provide a generic list in a similar format.}

\item{humandata, mousedata}{fill in the file 'ConsonsusTOM block. 1. rda' of human or mouse, which includes the variable 'consTomDS'.}

\item{humannet, mousenet}{fill in the file 'net. rda' of human or mouse, which includes the variable 'net'.'consTomDS' and 'net' were generated during the execution of WGCNA.

A code example for generating such files is:

net=blockwiseConsensusModules(multiExpr, blocks = NULL,}

\item{humanmodcolor, mousemodcolor}{The color of the module which you analyzed.}

\item{k}{(=20) The multiple relationship between 'background adjacency' and 'important adjacency' can be adjusted by k.}

\item{linecoloralpha}{(=c(0.40,0.40,0.40,0.40)) The transparency of the lines in the image can be specified separately by adjusting the linecoloralpha, which is crucial for visualizing the results}

\item{mosttimes}{Specify the maximum range for searching for k values, default to 500}

\item{quitname}{output file name}
}
\value{
4 picture of net compare will be build in
}
\description{
This package provides a WGCNA net cross species comparison method for an unspecified subset of genes. You need to provide WGCNA results for two species, including 'ConsonsusTOM block. 1. rda' and 'net'. Return four easily understandable visual network diagrams.
This project is based on cross species comparison between humans and mice, so the two species to be compared are referred to as humans and mice in this project. But we believe that this can also be directly applied to comparative studies of other species.
}
\examples{
nccs(humangene_1,mousegene_1,humandata_1,mousedata_1,humannet_1,mousenet_1,humanmodcolor_1,mousemodcolor_1,k=20,linecoloralpha=c(0.4,0.3,0.4,0.3),mosttimes=500,quitname='result of nccsWGCNA',means=c("Human Transformational Adjacency Network","Mouse Transformational Adjacency Network","Human-Mouse Transformational Adjacency Network","Mouse-Human Transformational Adjacency Network"))
}
