% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccsWGCNA-package.R
\docType{package}
\name{nccsWGCNA-package}
\alias{nccsWGCNA}
\alias{nccsWGCNA-package}
\title{nccsWGCNA: Net Compare Cross Species of WGCNA}
\description{
This package provides a WGCNA net cross species comparison method for an unspecified subset of genes. You need to provide WGCNA results for two species, including 'ConsonsusTOM block. 1. rda' and 'net'. Return four easily understandable visual network diagrams.This project is based on cross species comparison between humans and mice, so the two species to be compared are referred to as humans and mice in this project. But we believe that this can also be directly applied to comparative studies of other species.
}
\keyword{internal}
