% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpperExpansion.R
\name{UpperExpansion}
\alias{UpperExpansion}
\title{UpperExpansion}
\usage{
UpperExpansion(matrix1)
}
\arguments{
\item{matrix1}{A square matrix}
}
\value{
a list of value in the upper right corner of the matrix
}
\description{
a little function to take out upper triangular matrix expansion list
}
\examples{
UpperExpansion(matrix(2,nr=3,nc=3))
}
