% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccs.R
\name{nccs}
\alias{nccs}
\title{Net Compare Cross Species of WGCNA}
\usage{
nccs(
  humangene,
  mousegene,
  humandata,
  mousedata,
  humannet,
  mousenet,
  humanmodcolor,
  mousemodcolor,
  k = 20,
  linecoloralpha = c(0.4, 0.4, 0.4, 0.4)
)
}
\arguments{
\item{humangene, mousegene}{must be a one-to-one corresponding gene list to be analyzed.}

\item{humandata, mousedata}{fill in the file 'ConsonsusTOM block. 1. rda' of human or mouse, which includes the variable 'consTomDS'.}

\item{humannet, mousenet}{fill in the file 'net. rda' of human or mouse, which includes the variable 'net'.'consTomDS' and 'net' were generated during the execution of WGCNA.

A code example for generating such files is:

net=blockwiseConsensusModules(multiExpr, blocks = NULL,}

\item{humanmodcolor, mousemodcolor}{The color of the module which you analyzed.}

\item{k}{(=20) The multiple relationship between 'background adjacency' and 'important adjacency' can be adjusted by k.}

\item{linecoloralpha}{(=c(0.40,0.40,0.40,0.40)) The transparency of the lines in the image can be specified separately by adjusting the linecoloralpha, which is crucial for visualizing the results}
}
\value{
4 picture of net compare will be build in
}
\description{
This package provides a WGCNA net cross species comparison method for an unspecified subset of genes. You need to provide WGCNA results for two species, including 'ConsonsusTOM block. 1. rda' and 'net'. Return four easily understandable visual network diagrams.
This project is based on cross species comparison between humans and mice, so the two species to be compared are referred to as humans and mice in this project. But we believe that this can also be directly applied to comparative studies of other species.
}
\examples{
nccs(humangene_1,mousegene_1,humandata_1,mousedata_1,humannet_1,mousenet_1,humanmodcolor_1,mousemodcolor_1,k=20,linecoloralpha=c(0.4,0.3,0.4,0.3))
}
